PDF to Audio App/requirements.txt
pdfminer.six==20221105
gTTS==2.3.2